// apps/api/src/common/errors/success-code.ts
export enum SuccessCode {
  OPERACAO_SUCESSO = 'OPERACAO_SUCESSO',
  CONTA_CRIADA = 'CONTA_CRIADA',
  EMAIL_VERIFICADO = 'EMAIL_VERIFICADO',
  RECUPERACAO_SENHA_SOLICITADA = 'RECUPERACAO_SENHA_SOLICITADA',
  SENHA_REDEFINIDA = 'SENHA_REDEFINIDA',
  LOGIN_REALIZADO = 'LOGIN_REALIZADO',
  TOKEN_RENOVADO = 'TOKEN_RENOVADO',
  LOGOUT_REALIZADO = 'LOGOUT_REALIZADO',
  LAR_CRIADO = 'LAR_CRIADO',
  LARES_LISTADOS = 'LARES_LISTADOS',
  CONVITE_CRIADO = 'CONVITE_CRIADO',
  CONVITE_ACEITO = 'CONVITE_ACEITO',
  USUARIO_CARREGADO = 'USUARIO_CARREGADO',
  PLANO_CARREGADO = 'PLANO_CARREGADO',
  PLANO_ATUALIZADO = 'PLANO_ATUALIZADO',
  DOCUMENTO_CRIADO = 'DOCUMENTO_CRIADO',
  DOCUMENTOS_LISTADOS = 'DOCUMENTOS_LISTADOS',
  MEMBROS_LAR_LISTADOS = 'MEMBROS_LAR_LISTADOS',
  EVENTO_CRIADO = 'EVENTO_CRIADO',
  EVENTOS_LISTADOS = 'EVENTOS_LISTADOS',
}